# [Perf] Optimize `reshape_and_cache_flash` CUDA Kernel